Added Audio Manager and audio as well as specialised win/loss handlers. Also cleaned up a few scripts

+++ Space Invaders SFML/EntityManager.h
@@ -16,7 +16,7 @@ public:
 	virtual bool collisions();
 	virtual void draw(sf::RenderWindow &window);
 	virtual sf::Sprite getSprite() { return this->Sprite; };
-	virtual void setSpeed(int speed) { moveSpeed = speed; };
+	virtual void setSpeed(float speed) { moveSpeed = speed; };
 
 protected:
 	sf::Sprite Sprite;
